+++ src/main/java/pvpmode/api/server/overrides/PvPOverrideCondition.java
@@ -1,6 +1,8 @@
 package pvpmode.api.server.overrides;
 
 import net.minecraft.entity.player.EntityPlayer;
+import pvpmode.api.common.EnumPvPMode;
+import pvpmode.api.common.overrides.EnumForcedPvPMode;
 
 /**
  * The interface for an override condition. Implement it to create custom
@@ -25,23 +27,22 @@ public interface PvPOverrideCondition
      * <br/>
      * It also can return null, which means, that this condition doesn't apply.
      */
-    public Boolean isPvPEnabled (EntityPlayer player);
+    public EnumForcedPvPMode getForcedPvPMode (EntityPlayer player);
 
     /**
-     * Returns a message displayed to all users for the supplied player with the
-     * supplied forced mode when this player is forced into the supplied mode.
-     * Return null to display no message.
+     * Returns a message displayed when the supplied player is forced to the
+     * specified PvP Mode. Return null to display no message. If global is true, the
+     * message is displayed to all players, if false, it only will be displayed to
+     * the supplied player.
+     * 
+     * @param player
+     *            The player which PvP Mode will be forced
+     * @param forcedMode
+     *            The forced PvP Mode
+     * @param global
+     *            Whether the message will be sent to all players
+     * @return The message
      */
-    public String getForcedOverrideMessage (EntityPlayer player, Boolean mode);
-
-    /**
-     * Returns a message displayed to the supplied player with the supplied forced
-     * mode when this player is forced into the supplied mode. Return null to
-     * display no message.
-     */
-    public default String getLocalForcedOverrideMessage (EntityPlayer player, Boolean mode)
-    {
-        return getForcedOverrideMessage (player, mode);
-    }
+    public String getForcedOverrideMessage (EntityPlayer player, EnumPvPMode forcedMode, boolean global);
 
 }